pyproject.toml
[tool.poetry]
name = "name-classifier-rnn"
version = "0.1.0"
description = ""
authors = ["AshleyAlexJacob <alexjacob260@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.9"
#cpu
#torch = "^2.2.0"
# gpu
torch = {version = "2.2.0+cu121", source = "torch"}
matplotlib = "^3.8.2"
dagshub = "^0.3.13"
mlflow = "^2.10.2"
python-box = "^7.1.1"
pyyaml = "^6.0.1"
types-pyyaml = "^6.0.12.12"
python-dotenv = "^1.0.1"


[[tool.poetry.source]]
name = "PyPI"
priority = "primary"


[[tool.poetry.source]]
name = "torch"
url = "https://download.pytorch.org/whl/cu121"
priority = "supplemental"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
